Rename: dto.event.bookmark로 북마크 관련 dto 이동
